added more platforms, split up the ground into multiple pieces

+++ src/staticRect.h
@@ -6,9 +6,9 @@
 
 class StaticRect {
   b2Body* body;
-  int hw, hh; //half-width, half-height in meters
+  float hw, hh; //half-width, half-height in meters
 public:
-  StaticRect(int w, int h, float x, float y, b2World* world); //meters
+  StaticRect(float w, float h, float x, float y, b2World* world); //meters
 
   bool isInView(float absCameraPos); //pixels
   void draw(float absCameraPos); //pixels
